+++ src/main/java/net/javaguides/banking/dto/TransactionDTO.java
@@ -1,9 +1,10 @@
 package net.javaguides.banking.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public record TransactionDTO( Long id,
-                              Long accountId,
-                              Double amount,
-                              String transactionType,
-                              LocalDateTime timestamp) {}
+public record TransactionDTO(Long id,
+                             Long accountId,
+                             BigDecimal amount,
+                             String transactionType,
+                             LocalDateTime timestamp) {}
