+++ src/main/java/net/javaguides/banking/dto/TransferFundDTO.java
@@ -1,6 +1,8 @@
 package net.javaguides.banking.dto;
 
+import java.math.BigDecimal;
+
 public record TransferFundDTO(Long fromAccountId,
                               Long toAccountId,
-                              Double amount) {
+                              BigDecimal amount) {
 }
